Refactor iwp into multiple files
