+++ llarp/CMakeLists.txt
@@ -166,7 +166,8 @@ set(LIB_SRC
   handlers/exit.cpp
   handlers/null.cpp
   handlers/tun.cpp
-  link/iwp_internal.cpp
+  link/iwp_linklayer.cpp
+  link/iwp_outermessage.cpp
   link/iwp.cpp
   link/server.cpp
   link/session.cpp
